Wrap Clutter::Effect animation API
Clutter::Effect is a though beast to bind. I've chosen the form of class
methods under the Clutter::Effect package name, which is the form that
most roughly matches the C library usage.

The Clutter::Effect->move() method is not yet wrapped because it accepts
a pointer to a static array of knots. This will probably need changing
upstream, because we don't have a convenient append_knot() method here
for dynamically allocated knots.