Bind ClutterCairo integration library
clutter-cairo is an integration library providing a single actor, ClutterCairo.
The Clutter Perl bindings wrap this actor as Clutter::Texture::Cairo, following
the naming convention of Clutter::Texture::Clone.

If the clutter-cairo development files are found, the Clutter::Texture::Cairo
actor is automatically built and exported under the Clutter namespace.

With Clutter::Texture::Cairo you can get a Cairo::Context and draw on the actor
with the Cairo drawin primitives (you need the Cairo Perl bindings in order to
do this).