Add list-style accessors to ClutterBehaviourEllipse
The ellipse behaviour has far too many accessors already, but lacks methods
to set similar properties in one call. This patch adds a set_angles() method
for setting the initial and final angles and a set_size() method for setting
the width and height of the ellipsis; relative accessors are also provided.