Allow overriding ClutterBox::pack_child() and ClutterBox::unpack_child()
Provide a way to wrap and unwrap a ClutterBoxChild as an arrayref or as
an hashref, and use it to make the Perl wrappers around the ::pack_child()
and ::unpack_child() sane.

Change the return value of Clutter::Box::query_child() and
Clutter::Box::query_nth_child() to an hashref containing the ClutterBoxChild
members as keys.

Exercise the API in the boxes.pl example.