Fix code reference for built-in alpha functions
Apparently when we switched to using closures something
broke inside the Clutter::Alpha bindings and we could not
set a built-in alpha function like this:

  \&Clutter::Alpha::ramp_inc

anymore, which required wrapping every alpha function using
a local code reference like:

  sub ramp_inc {
      return Clutter::Alpha::ramp_inc($_[0]);
  }

and it surely is not at all optimal.

Thanks to Andrew Ruthven for spotting this.