[cogl] Fix typos in the cogl_frustum() bindings
The Cogl::frustum implementation suffered from an acute case of
cut-and-pastitis, a very serious disease that affects function calls
with more than three arguments all of the same time requiring the
same lengthy macro.

The only way to cure a cut-and-pastitis is to install a brain in the
developer that wrote the bindings.