[test] Check alpha value when advancing the timeline
The changes made in Clutter 0.8.2 to ClutterAlpha::get_alpha() make
it possible to test that advancing the timeline bound to the alpha
instance we get a different alpha value out of it.

In this case, we can check that an alpha is working correctly even
though we don't have a spinning main loop, by manually advancing the
timeline to the last frame and verifying that the alpha value matches
the Clutter::Alpha->MAX_ALPHA constant.