Allow using normalized values for CoglColors
CoglColor in C allows storing a color using bytes for each channel
and normalized values between 0 and 1.

If the first element of the ARRAY that we want to convert to a
CoglColor is an integer, we use bytes; if it is a float, then we
use normalized values.

When we convert a CoglColor to Perl, instead, we always use
normalized values and let COGL do the right thing when it needs
to handle the color internally.