Rework Clutter::PathNode bindings
The PathNode structure is very similar to the CairoPath components,
so the bindings should actually match. For this reason, a PathNode
is now defined as a hash reference with two keys:

        - type   => any('move-to', 'line-to', 'curve-to', 'close'),
        - points => [ ]

The number elements inside the array reference for 'points' depend
on the 'type' key:

        [ [ x, y ] ]                            : move-to, line-to
        [ [ x1, y1 ], [ x2, y2 ], [ x3, y3 ] ]  : curve-to
        [ ]                                     : close

Except for the mismatched 'close'/'close-path', this is exactly
like a Cairo::Path.