Move the old ChangeLog out of the dist
The ChangeLog file was used for Subversion and it was distributed
inside the tarball. Since we switched to Git and moved the changes
logging to the commit messages, there's no reason for us to have a
ChangeLog file -- thus ChangeLog was renamed ChangeLog.SVN and
maintained in the repository purely for historical reasons.

Having the ChangeLog in the dist package, though, was just redundant:
if you are using a package then you'll probably just want to use the
NEWS file, since you don't care about the history of the project, or
you'd be using a clone of the Git repository.

For this reason, we can safely remove the ChangeLog.SVN file from the
dist and leave it in the repository.