[container] Allow omitting FOREACH_WITH_INTERNALS
The FOREACH_WITH_INTERNALS virtual function should be optional, since
it can be completely substituted by FOREACH for simple containers. This
is the behaviour of the upstream C API anyway, but since Perl bindings
must override the whole interface, the upstream check for a NULL
::foreach_with_internals() function pointer fails.

We just need a variant of the GET_METHOD() macro that checks for a
fallback method.